fix: console table when wrong command is sent
