+++ commands/generate.command.ts
@@ -1,11 +1,8 @@
 import * as chalk from 'chalk';
-import * as Table from 'cli-table3';
 import { Command, CommanderStatic } from 'commander';
-import { AbstractCollection, CollectionFactory } from '../lib/schematics';
-import { Schematic } from '../lib/schematics/nest.collection';
-import { loadConfiguration } from '../lib/utils/load-configuration';
 import { AbstractCommand } from './abstract.command';
 import { Input } from './command.input';
+import { getCollection, getSchematics, buildSchematicsListAsTable } from '../lib/ui/table';
 
 export class GenerateCommand extends AbstractCommand {
   public async load(program: CommanderStatic): Promise<void> {
@@ -104,47 +101,11 @@ export class GenerateCommand extends AbstractCommand {
   }
 
   private async buildDescription(): Promise<string> {
-    const collection = await this.getCollection();
+    const collection = await getCollection();
     return (
       'Generate a Nest element.\n' +
       `  Schematics available on ${chalk.bold(collection)} collection:\n` +
-      this.buildSchematicsListAsTable(await this.getSchematics(collection))
+      buildSchematicsListAsTable(await getSchematics(collection, 'schematic'))
     );
-  }
-
-  private buildSchematicsListAsTable(schematics: Schematic[]): Promise<string> {
-    const leftMargin = '    ';
-    const tableConfig = {
-      head: ['name', 'alias', 'description'],
-      chars: {
-        'left': leftMargin.concat('│'),
-        'top-left': leftMargin.concat('┌'),
-        'bottom-left': leftMargin.concat('└'),
-        'mid': '',
-        'left-mid': '',
-        'mid-mid': '',
-        'right-mid': '',
-      },
-    };
-    const table: any = new Table(tableConfig);
-    for (const schematic of schematics) {
-      table.push([
-        chalk.green(schematic.name),
-        chalk.cyan(schematic.alias),
-        schematic.description,
-      ]);
-    }
-    return table.toString();
-  }
-
-  private async getCollection(): Promise<string> {
-    const configuration = await loadConfiguration();
-    return configuration.collection;
-  }
-
-  private async getSchematics(collection: string): Promise<Schematic[]> {
-    const abstractCollection: AbstractCollection =
-      CollectionFactory.create(collection);
-    return abstractCollection.getSchematics('schematic');
   }
 }
